Seems to be working! But sheesh.. lot to unpack here
The biggest mistake I made was logging retrieved rather than results -- which then inspired tearing down the allowList in searchByVector to try to debug the filter == neighborFilter thing at search time. Anyways, what I rebuilt fixed it but it's not clear if I needed to do that to begin with.