Testing, not looking good
The key hyperparameters are (1) K, if a node has less than K neighbors we add additional edges, and (2) M, the # of additional edges added.

Found the best results with K = 5 and M = 128. But still a very small boost in latency.

Pushing this and adding test of connecting majority nodes that don't have any edges to minority nodes.